requirements.txt
astroid==2.1.0
attrs==19.3.0
azureml-dataprep==0.5.3
azureml-dataprep-native==11.1.2
backcall==0.1.0
bleach==3.1.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
distro==1.3.0
dotnetcore2==2.1.7
entrypoints==0.3
idna==2.8
ipykernel==5.1.3
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.4
jedi==0.15.1
Jinja2==2.10.1
json5==0.8.5
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
jupyterlab==1.2.5
jupyterlab-server==1.0.6
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.15.4
pandas==0.23.4
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
plotly==3.4.2
prometheus-client==0.7.1
prompt-toolkit==2.0.9
Pygments==2.4.2
pylint==2.2.2
pyparsing==2.3.0
pyrsistent==0.15.4
pytest==5.0.1
python-dateutil==2.8.0
pytz==2018.7
pywin32==225
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.21.0
retrying==1.3.3
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
terminado==0.8.2
testpath==0.4.2
tornado==6.0.3
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.10.11
